Add function docstrings and reorganize functions in separate file
